zenflow 빌드 에러 3건 수정 — 타입 리졸버 KEY fallback 추가
- SSaC 타입 리졸버: VALUE 기반 DDL 조회 실패 시 KEY 기반 DDL fallback 추가
- zenflow func spec: CheckCreditsRequest.Balance int32→int64 (fullend DDL 매핑 일치)
- zenflow SSaC: UpdateCredits Amount→CreditsBalance (DDL 컬럼명 일치)
- gigbridge/zenflow 산출물 재생성, 양쪽 go build 성공

+++ artifacts/zenflow/backend/internal/billing/check_credits.go
@@ -7,7 +7,7 @@ import "fmt"
 // @description Validates that the organization has sufficient credits
 
 type CheckCreditsRequest struct {
-	Balance int32
+	Balance int64
 }
 
 type CheckCreditsResponse struct{}
